Build from git
The Dolphin project is sadly one of those that never makes stable
releases and as such it is multiple years behind what the current
project actually is and users want.

Along with updating this also switches to the new Qt interface.
It is still actively in devleopment but it approaching completion
and will become the default UI soon. This switch is happening
early here simply because WxWidgets just brings more bugs and build
complexity with it than its worth.